Fix clicking to other pages through task queue
You can't call click from the browser and have Puppeteer navigate the
window while the browser's JavaScript executor is waiting for the
promise to be resolved, OR have the browser navigate on its own.
This is because JavaScript runs on a page-by-page basis in the browser,
and when you change the page if your JavaScript is running it will
break. Therefore, we implement a task queue that allows us to process
the click after the JavaScript execution is over.